refactor database interface and implementation into db subpackage
We now have:
- db/dbi.go -- interface, no dependency on any particular database
- db/sqlite/... -- sqlite implementation, depends on dbi
- the rendezvous.NewRendezvousService API has been redefined to NewRendezvouServiceWithDB
  service importer constructs the database of choice
- the rendezvous package has no dependences on db implementation, so clients can import
  without linking SQLite in the binary
