+++ db/dbi.go
@@ -1,0 +1,21 @@
+package dbi
+
+import (
+	peer "github.com/libp2p/go-libp2p-peer"
+)
+
+type RegistrationRecord struct {
+	Id    peer.ID
+	Addrs [][]byte
+	Ns    string
+	Ttl   int
+}
+
+type DB interface {
+	Close() error
+	Register(p peer.ID, ns string, addrs [][]byte, ttl int) error
+	Unregister(p peer.ID, ns string) error
+	CountRegistrations(p peer.ID) (int, error)
+	Discover(ns string, cookie []byte, limit int) ([]RegistrationRecord, []byte, error)
+	ValidCookie(ns string, cookie []byte) bool
+}
